move more complex signers to own files (#51266)
